examples/demo/hyperpocket-demo/pyproject.toml
[project]
name = "hyperpocket-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "langchain>=0.3.14",
    "langchain-openai>=0.3.0",
    "hyperpocket",
    "hyperpocket-langchain",
    "hyperpocket-langgraph",
    "hyperpocket-llamaindex",
    "hyperpocket-openai",
    "praw>=7.8.1",
    "pydantic>=2.10.1",
    "pydantic-core>=2.27.1",
    "slack-sdk==3.33.4",
    "typing-extensions==4.12.2",
]

[tool.uv.sources]
hyperpocket = { path = "../../../libs/hyperpocket", editable = true }
hyperpocket-langchain = { path = "../../../libs/extensions/langchain", editable = true }
hyperpocket-langgraph = { path = "../../../libs/extensions/langgraph", editable = true }
hyperpocket-llamaindex = { path = "../../../libs/extensions/llamaindex", editable = true }
hyperpocket-openai = { path = "../../../libs/extensions/openai", editable = true }
